refactor: add all triggers in one place
