+++ packages/pieces/community/krisp-call/src/index.ts
@@ -4,9 +4,7 @@ import { addContact } from './lib/actions/add-contact';
 import { deleteContacts } from './lib/actions/delete-contacts';
 import { sendSms } from './lib/actions/send-sms';
 import { sendMms } from './lib/actions/send-mms';
-import { newContact } from './lib/triggers/new-contact';
-import { newMms } from './lib/triggers/new-mms';
-import { newVoicemail } from './lib/triggers/new-voicemail';
+import { triggers } from './lib/triggers';
 
 export const krispcallAuth = PieceAuth.CustomAuth({
   props: {
@@ -40,9 +38,8 @@ export const KrispCall = createPiece({
   displayName: 'Krispcall',
   auth: krispcallAuth,
   minimumSupportedRelease: '0.20.0',
-  logoUrl:
-    'https://krispcall.com/wp-content/uploads/2023/06/krispcall-favicon.svg',
+  logoUrl: 'https://cdn.activepieces.com/pieces/krispcall.svg',
   authors: [],
   actions: [addContact, deleteContacts, sendSms, sendMms],
-  triggers: [newContact, newMms, newVoicemail],
+  triggers: triggers,
 });
